add a basic interactive plot of ssts
This plot runs in matplotlib and could use some enhancement. It allows access to all important parameters to demonstrate behavior. An ideal implementation would reduce the need for many of these to be accessed (i.e. percentages and “x”-location would be automatically determined based on the “y” values). A second plot showing this “automatic” behavior should be added in a future commit.